Add files via upload
Fixing an issue that was fixed before the due date and was overwritten, just saw it right now. If you look back at the commits, it fixes a critical error on line 169 with some smaller aesthetic changes, then is overwritten in another commit. This commit basically fixes what didn't need to be changed in the commit that got rid of everything.